wip cleanup with of philip_pal with\ mm_pal
